chore(orchestrator): promote NFS plugins from /alpha to stable
Fixes: https://redhat.atlassian.net/browse/RHIDP-14417
