+++ workspaces/orchestrator/docs/workflows-tab.md
@@ -66,7 +66,7 @@ To expose workflows directly on the entity page, update
 import {
   OrchestratorCatalogTab,
   IsOrchestratorCatalogTabAvailable,
-} from '@red-hat-developer-hub/backstage-plugin-orchestrator';
+} from '@red-hat-developer-hub/backstage-plugin-orchestrator/legacy';
 
 const serviceEntityPage = (
   <EntityLayout>
