+++ workspaces/orchestrator/packages/app-legacy/src/App.tsx
@@ -46,8 +46,8 @@ import {
 import { ReportIssue } from '@backstage/plugin-techdocs-module-addons-contrib';
 import { TechDocsAddons } from '@backstage/plugin-techdocs-react';
 import { UserSettingsPage } from '@backstage/plugin-user-settings';
-import { OrchestratorPage } from '@red-hat-developer-hub/backstage-plugin-orchestrator';
-import { orchestratorFormWidgetsPlugin } from '@red-hat-developer-hub/backstage-plugin-orchestrator-form-widgets';
+import { OrchestratorPage } from '@red-hat-developer-hub/backstage-plugin-orchestrator/legacy';
+import { orchestratorFormWidgetsPlugin } from '@red-hat-developer-hub/backstage-plugin-orchestrator-form-widgets/legacy';
 import { customAuthProviderPlugin } from 'custom-authentication-provider-module';
 import { getThemes } from '@red-hat-developer-hub/backstage-plugin-theme';
 import { NotificationsPage } from '@backstage/plugin-notifications';
@@ -58,7 +58,7 @@ import { apis } from './apis';
 import { entityPage } from './components/catalog/EntityPage';
 import { Root } from './components/Root';
 import { searchPage } from './components/search/SearchPage';
-import { orchestratorTranslations } from '@red-hat-developer-hub/backstage-plugin-orchestrator';
+import { orchestratorTranslations } from '@red-hat-developer-hub/backstage-plugin-orchestrator/legacy';
 
 const app = createApp({
   apis,
