--- workspaces/orchestrator/plugins/orchestrator-form-widgets/report-alpha.api.md
@@ -1,38 +1,0 @@
-## API Report File for "@red-hat-developer-hub/backstage-plugin-orchestrator-form-widgets"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-import { AnyApiFactory } from '@backstage/frontend-plugin-api';
-import { ApiFactory } from '@backstage/frontend-plugin-api';
-import { ExtensionBlueprintParams } from '@backstage/frontend-plugin-api';
-import { ExtensionDataRef } from '@backstage/frontend-plugin-api';
-import { OverridableExtensionDefinition } from '@backstage/frontend-plugin-api';
-import { OverridableFrontendPlugin } from '@backstage/frontend-plugin-api';
-
-// @alpha
-const _default: OverridableFrontendPlugin<
-  {},
-  {},
-  {
-    'api:orchestrator-form-widgets': OverridableExtensionDefinition<{
-      kind: 'api';
-      name: undefined;
-      config: {};
-      configInput: {};
-      output: ExtensionDataRef<AnyApiFactory, 'core.api.factory', {}>;
-      inputs: {};
-      params: <
-        TApi,
-        TImpl extends TApi,
-        TDeps extends { [name in string]: unknown },
-      >(
-        params: ApiFactory<TApi, TImpl, TDeps>,
-      ) => ExtensionBlueprintParams<AnyApiFactory>;
-    }>;
-  }
->;
-export default _default;
-
-// (No @packageDocumentation comment for this package)
-```
