+++ workspaces/orchestrator/plugins/orchestrator-form-widgets/report-legacy.api.md
@@ -1,0 +1,25 @@
+## API Report File for "@red-hat-developer-hub/backstage-plugin-orchestrator-form-widgets"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { ApiFactory } from '@backstage/frontend-plugin-api';
+import { BackstagePlugin } from '@backstage/core-plugin-api';
+import { OrchestratorFormApi } from '@red-hat-developer-hub/backstage-plugin-orchestrator-form-api';
+
+// Warning: (ae-forgotten-export) The symbol "FormWidgetsApi" needs to be exported by the entry point legacyExports.d.ts
+//
+// @public
+export const orchestratorFormApiFactory: ApiFactory<
+  OrchestratorFormApi,
+  FormWidgetsApi,
+  {
+    [x: string]: unknown;
+  }
+>;
+
+// @public
+export const orchestratorFormWidgetsPlugin: BackstagePlugin<{}, {}, {}>;
+
+// (No @packageDocumentation comment for this package)
+```
