+++ workspaces/orchestrator/plugins/orchestrator-form-widgets/report.api.md
@@ -3,23 +3,36 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
+import { AnyApiFactory } from '@backstage/frontend-plugin-api';
 import { ApiFactory } from '@backstage/frontend-plugin-api';
-import { BackstagePlugin } from '@backstage/core-plugin-api';
-import { OrchestratorFormApi } from '@red-hat-developer-hub/backstage-plugin-orchestrator-form-api';
+import { ExtensionBlueprintParams } from '@backstage/frontend-plugin-api';
+import { ExtensionDataRef } from '@backstage/frontend-plugin-api';
+import { OverridableExtensionDefinition } from '@backstage/frontend-plugin-api';
+import { OverridableFrontendPlugin } from '@backstage/frontend-plugin-api';
 
-// Warning: (ae-forgotten-export) The symbol "FormWidgetsApi" needs to be exported by the entry point index.d.ts
-//
 // @public
-export const orchestratorFormApiFactory: ApiFactory<
-  OrchestratorFormApi,
-  FormWidgetsApi,
+const _default: OverridableFrontendPlugin<
+  {},
+  {},
   {
-    [x: string]: unknown;
+    'api:orchestrator-form-widgets': OverridableExtensionDefinition<{
+      kind: 'api';
+      name: undefined;
+      config: {};
+      configInput: {};
+      output: ExtensionDataRef<AnyApiFactory, 'core.api.factory', {}>;
+      inputs: {};
+      params: <
+        TApi,
+        TImpl extends TApi,
+        TDeps extends { [name in string]: unknown },
+      >(
+        params: ApiFactory<TApi, TImpl, TDeps>,
+      ) => ExtensionBlueprintParams<AnyApiFactory>;
+    }>;
   }
 >;
-
-// @public
-export const orchestratorFormWidgetsPlugin: BackstagePlugin<{}, {}, {}>;
+export default _default;
 
 // (No @packageDocumentation comment for this package)
 ```
