+++ workspaces/orchestrator/plugins/orchestrator-form-widgets/src/legacyExports.ts
@@ -13,4 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { orchestratorTranslationsModule as default } from './index';
+
+export {
+  orchestratorFormWidgetsPlugin,
+  orchestratorFormApiFactory,
+} from './plugin';
