+++ workspaces/orchestrator/plugins/orchestrator/app-config.yaml
@@ -1,15 +1,19 @@
 dynamicPlugins:
   frontend:
     red-hat-developer-hub.backstage-plugin-orchestrator:
+      # Legacy exports require `module: Legacy` — they are not available on the default module.
       translationResources:
         - importName: orchestratorTranslations
           ref: orchestratorTranslationRef
+          module: Legacy
       appIcons:
         - name: orchestratorIcon
           importName: OrchestratorIcon
+          module: Legacy
       dynamicRoutes:
         - path: /orchestrator
           importName: OrchestratorPage
+          module: Legacy
           menuItem:
             icon: orchestratorIcon
             text: Orchestrator
@@ -25,6 +29,7 @@ dynamicPlugins:
       mountPoints:
         - mountPoint: entity.page.workflows/cards
           importName: OrchestratorCatalogTab
+          module: Legacy
           config:
             layout:
               gridColumn: '1 / -1'
