+++ workspaces/orchestrator/plugins/orchestrator/dev/index.tsx
@@ -26,12 +26,10 @@ import userSettingsPlugin from '@backstage/plugin-user-settings/alpha';
 // eslint-disable-next-line @backstage/no-ui-css-imports-in-non-frontend
 import '@backstage/ui/css/styles.css';
 
-import orchestratorFormWidgetsPlugin from '@red-hat-developer-hub/backstage-plugin-orchestrator-form-widgets/alpha';
+import orchestratorFormWidgetsPlugin from '@red-hat-developer-hub/backstage-plugin-orchestrator-form-widgets';
 import { rhdhThemeModule } from '@red-hat-developer-hub/backstage-plugin-theme/alpha';
 
-import orchestratorPlugin, {
-  orchestratorTranslationsModule,
-} from '../src/alpha';
+import orchestratorPlugin, { orchestratorTranslationsModule } from '../src';
 import { devNavModule } from './nav';
 
 const App = createApp({
