+++ workspaces/orchestrator/plugins/orchestrator/package.json
@@ -2,18 +2,22 @@
   "name": "@red-hat-developer-hub/backstage-plugin-orchestrator",
   "version": "5.10.0",
   "license": "Apache-2.0",
-  "main": "src/index.ts",
-  "types": "src/index.ts",
+  "main": "src/index.tsx",
+  "types": "src/index.tsx",
   "exports": {
-    ".": "./src/index.ts",
-    "./alpha": "./src/alpha.tsx",
+    ".": "./src/index.tsx",
+    "./alpha": "./src/alpha.ts",
+    "./legacy": "./src/legacyExports.ts",
     "./orchestrator-translations-module": "./src/orchestratorTranslationsModuleExport.ts",
     "./package.json": "./package.json"
   },
   "typesVersions": {
     "*": {
       "alpha": [
-        "src/alpha.tsx"
+        "src/alpha.ts"
+      ],
+      "legacy": [
+        "src/legacyExports.ts"
       ],
       "orchestrator-translations-module": [
         "src/orchestratorTranslationsModuleExport.ts"
@@ -135,8 +139,9 @@
   "scalprum": {
     "name": "red-hat-developer-hub.backstage-plugin-orchestrator",
     "exposedModules": {
-      "PluginRoot": "./src/index.ts",
-      "Alpha": "./src/alpha.tsx"
+      "PluginRoot": "./src/index.tsx",
+      "Legacy": "./src/legacyExports.ts",
+      "OrchestratorTranslationsModule": "./src/orchestratorTranslationsModuleExport.ts"
     }
   },
   "maintainers": [
